+++ data/example_tweets.tsv
@@ -1,5 +1,5 @@
-tweetid	text	urls
-714752339819757568	In @sciencemagazine: Rare variant found that raises HDL cholesterol and increases risk of coronary heart disease: https://t.co/4xpL3KlGy9	['https://www.science.org/doi/full/10.1126/science.aad3517']
-361218917303193600	Teach children to treat animals responsibly do not teach captivity! Join us http://t.co/UR15gQPatU #FreeAllCetacea via @FreeAllCetacea	['http://www.wdcs.org/']
-712710761240350720	Violence is a leading cause of death for Americans 10-24 yrs old. @NICHD_NIH research on youth violence prevention https://t.co/9ZD58zGUhI	['https://1.usa.gov/1q0MqOd/']
-741038054446600192	"“At a point now where I understand plant temperaments — how vulnerable they are, but also how strong they can become"" #Precision #AgTech"	[]
+text	urls
+In @sciencemagazine: Rare variant found that raises HDL cholesterol and increases risk of coronary heart disease: https://t.co/4xpL3KlGy9	['https://www.science.org/doi/full/10.1126/science.aad3517']
+Teach children to treat animals responsibly do not teach captivity! Join us http://t.co/UR15gQPatU #FreeAllCetacea via @FreeAllCetacea	['http://www.wdcs.org/']
+Violence is a leading cause of death for Americans 10-24 yrs old. @NICHD_NIH research on youth violence prevention https://t.co/9ZD58zGUhI	['https://1.usa.gov/1q0MqOd/']
+"“At a point now where I understand plant temperaments — how vulnerable they are, but also how strong they can become"" #Precision #AgTech"	[]
