+++ data/example_tweets_cat1_cat2_cat3.tsv
@@ -1,0 +1,5 @@
+text	is_claim	claim_sentence	has_sciterm	sciterms	is_cat1	urls	sci_subdomain	has_sci_subdomain	sci_mag_domain	has_sci_mag_domain	sci_news_domain	has_sci_news_domain	is_cat2	mentions_science_research_in_general	mentions_scientist	mentions_publications	mentions_research_method	is_cat3
+In @sciencemagazine: Rare variant found that raises HDL cholesterol and increases risk of coronary heart disease: https://t.co/4xpL3KlGy9	True	in @sciencemagazine: rare variant found that raises hdl cholesterol and increases risk of coronary heart disease: https://t.co/4xpl3klgy9	True	['cholesterol', 'disease', 'heart', 'risk', 'variant']	True	['https://www.science.org/doi/full/10.1126/science.aad3517']	www.science.org	True	www.science.org	True		False	True					False
+Teach children to treat animals responsibly do not teach captivity! Join us http://t.co/UR15gQPatU #FreeAllCetacea via @FreeAllCetacea	True	teach children to treat animals responsibly do not teach captivity!	True	['animals']	True	['http://www.wdcs.org/']		False		False		False	False					False
+Violence is a leading cause of death for Americans 10-24 yrs old. @NICHD_NIH research on youth violence prevention https://t.co/9ZD58zGUhI	True	violence is a leading cause of death for americans 10-24 yrs old.	True	['prevention', 'research']	True	['https://1.usa.gov/1q0MqOd/']		False		False		False	False	research on				True
+"“At a point now where I understand plant temperaments — how vulnerable they are, but also how strong they can become"" #Precision #AgTech"	False		True	['plant']	False	[]	[]	False	[]	False	[]	False	False					False
