--- data/example_tweets_cat1_heuristics.tsv
@@ -1,5 +1,0 @@
-tweetid	text	urls	is_claim	claim_sentence	has_sciterm	sciterms	is_cat1
-714752339819757568	In @sciencemagazine: Rare variant found that raises HDL cholesterol and increases risk of coronary heart disease: https://t.co/4xpL3KlGy9	['https://www.science.org/doi/full/10.1126/science.aad3517']	True	in @sciencemagazine: rare variant found that raises hdl cholesterol and increases risk of coronary heart disease: https://t.co/4xpl3klgy9	True	['cholesterol', 'disease', 'heart', 'risk', 'variant']	True
-361218917303193600	Teach children to treat animals responsibly do not teach captivity! Join us http://t.co/UR15gQPatU #FreeAllCetacea via @FreeAllCetacea	['http://www.wdcs.org/']	True	teach children to treat animals responsibly do not teach captivity!	True	['animals']	True
-712710761240350720	Violence is a leading cause of death for Americans 10-24 yrs old. @NICHD_NIH research on youth violence prevention https://t.co/9ZD58zGUhI	['https://1.usa.gov/1q0MqOd/']	True	violence is a leading cause of death for americans 10-24 yrs old.	True	['prevention', 'research']	True
-741038054446600192	"“At a point now where I understand plant temperaments — how vulnerable they are, but also how strong they can become"" #Precision #AgTech"	[]	False		True	['plant']	False
