+++ src/lists/discovery_verbs_kws.txt
@@ -1,0 +1,30 @@
+predict
+discover
+say
+find
+show
+develop
+research
+highlight
+constitute
+suggest
+indicate
+demonstrate
+show
+reveal
+provide
+illustrate
+describe
+conclude
+support
+establish
+propose
+advocate
+determine
+confirm
+argue
+impl
+display
+offer
+underline
+allow
