+++ src/lists/exclude_methods.txt
@@ -1,0 +1,10 @@
+life story
+sample size
+ideal type
+life experiences
+data privacy
+role playing
+data quality
+survey results
+research grants
+false negative
