+++ src/lists/false_scientific_words.txt
@@ -1,0 +1,502 @@
+haven
+practice
+fans
+call
+place
+special
+account
+young
+small
+point
+games
+act
+join
+past
+words
+events
+respect
+simple
+current
+student
+major
+account
+cover
+moment
+direct
+complete
+five
+cute
+club
+non
+final
+voice
+essential
+date
+tour
+ads
+till
+card
+youtube
+write
+pitch
+interviews
+duo
+cant
+even
+community
+force
+support
+even
+food
+associated
+death
+say
+author
+find
+reading
+being
+work
+air
+storms
+security
+real
+say
+kissing
+day
+need
+support
+feedback
+being
+perfect
+religious
+curriculum
+good
+line
+schools
+women
+link
+meet
+mind
+eyes
+crew
+support
+being
+work
+ago
+load
+associated
+clean
+first
+insurance
+year
+end
+good
+lie
+support
+nah
+rain
+good
+share
+time
+support
+support
+bird
+bus
+grade
+groups
+good
+linked
+negative
+self
+number
+real
+support
+satellites
+tap
+wire
+world
+support
+problem
+range
+resource
+support
+day
+support
+trade
+almost
+authority
+symptoms
+women
+strength
+support
+centre
+footprint
+sets
+groups
+higher
+resolution
+series
+volatile
+fly
+support
+minor
+first
+time
+tax
+game
+associated
+good
+year
+markets
+symptoms
+driver
+lesson
+support
+women
+support
+support
+world
+sink
+time
+change
+cost
+green
+scope
+speed
+key
+angles
+end
+scheme
+free
+support
+message
+carrier
+communications
+support
+package
+cactus
+fruit
+being
+full
+support
+strong
+action
+good
+free
+hotspot
+internet
+support
+local
+support
+entertainment
+bit
+groups
+support
+need
+value
+emotions
+negative
+night
+support
+complex
+freezing
+province
+rain
+snow
+killing
+plan
+global
+kind
+movement
+message
+support
+schools
+peace
+prayer
+pressure
+cars
+support
+demand
+food
+free
+effectiveness
+food
+medications
+design
+ear
+vibration
+debt
+service
+head
+school
+line
+local
+internal
+support
+reason
+web
+negative
+find
+higher
+zoning
+collection
+compliance
+catastrophic
+control
+stand
+support
+learning
+mental
+students
+support
+entity
+laws
+even
+process
+year
+family
+spread
+support
+understanding
+want
+positive
+good
+note
+time
+top
+being
+activity
+higher
+active
+burn
+change
+color
+red
+aging
+being
+fix
+processes
+eventually
+label
+energy
+need
+support
+time
+change
+journey
+related
+support
+fire
+stream
+being
+expression
+free
+speaking
+group
+member
+support
+class
+classes
+community
+game
+grade
+play
+school
+second
+skills
+students
+time
+matter
+world
+child
+greedy
+sound
+support
+interaction
+leadership
+member
+position
+blocking
+need
+reason
+support
+case
+list
+community
+depreciation
+assembly
+support
+total
+behaviors
+change
+negative
+positive
+higher
+world
+year
+industry
+support
+activists
+century
+food
+critical
+mental
+positive
+thinking
+thought
+energy
+support
+country
+site
+group
+front
+support
+energy
+higher
+independence
+prices
+industry
+year
+sign
+face
+problem
+support
+world
+want
+information
+matrix
+power
+prayer
+reason
+say
+world
+back
+city
+energy
+habit
+support
+hair
+event
+first
+link
+second
+support
+child
+governor
+support
+age
+children
+game
+groups
+reverse
+training
+touch
+back
+culture
+good
+need
+real
+time
+annual
+spring
+real
+women
+need
+relief
+support
+gesture
+support
+enough
+harvesting
+interference
+enough
+household
+need
+range
+discussion
+information
+patent
+support
+data
+group
+groups
+lots
+span
+year
+being
+list
+top
+landslide
+support
+bit
+killing
+channel
+end
+ran
+standards
+revolution
+absolutely
+entire
+need
+thread
+digital
+support
+back
+situation
+breaking
+industry
+path
+support
+music
+participation
+similar
+face
+support
+role
+product
+big
+primary
+significant
+support
+mad
+being
+path
+link
+group
+handle
+need
+school
+students
+support
+teachers
+country
+day
+source
+vital
+class
+want
+wealth
+work
+end
+host
+focus
+root
+support
+time
+world
+students
+support
+add
+bar
+bit
+couple
+film
+set
+star
+support
+motor
+city
+support
+cut
+finds
+play
