+++ src/lists/predicates.txt
