+++ src/lists/publications_kws.txt
@@ -1,0 +1,10 @@
+publications
+posters
+reports
+statistics
+datasets
+findings
+papers
+studies
+experiments
+surveys
