+++ src/lists/sc_methods.txt
