+++ src/lists/science_mags_domains.txt
@@ -1,4 +1,3 @@
-domain
 sciencedaily.com
 newscientist.com
 scientificamerican.com
