+++ src/lists/science_research_in_general_kws.txt
@@ -1,0 +1,15 @@
+research on
+research in
+research for
+research from
+research of
+research to
+research at
+research by
+research
+science of
+science to
+science
+sciences of
+sciences to
+sciences
