+++ src/lists/scientists_kws.txt
@@ -1,0 +1,12 @@
+research team
+research group
+scientist
+researcher
+psychologist
+chemist
+physician
+biologist
+economist
+engineer
+physicist
+geologist
