+++ src/lists/text_speak_words.txt
@@ -1,0 +1,17 @@
+lol
+lmao
+lmfao
+omg
+ppl
+pls
+thx
+sry
+dah
+dat
+soo
+isn
+stan
+abt
+dawg
+cha
+jk
